+++ src/node.h
@@ -2522,13 +2522,6 @@ public:
 
   bool isMethod() const { return this->kind == FUNC && static_cast<bool>(this->getRecvTypeNode()); }
 
-  /**
-   * for CLI record
-   * @return
-   * if not found, return empty string
-   */
-  StringRef getCLIName() const;
-
   void dump(NodeDumper &dumper) const override;
 };